fix NPE when ItemBridge not loaded
